// frontend/src/api/model/pageInfoUser.ts
/**
 * Generated by orval v8.4.0 🍺
 * Do not edit manually.
 * Movie Review System API
 * Backend API for Movie Review Website
 * OpenAPI spec version: v1.0
 */
import type { User } from "./user";

export interface PageInfoUser {
  total?: number;
  list?: User[];
  pageNum?: number;
  pageSize?: number;
  size?: number;
  startRow?: number;
  endRow?: number;
  pages?: number;
  prePage?: number;
  nextPage?: number;
  isFirstPage?: boolean;
  isLastPage?: boolean;
  hasPreviousPage?: boolean;
  hasNextPage?: boolean;
  navigatePages?: number;
  navigatepageNums?: number[];
  navigateFirstPage?: number;
  navigateLastPage?: number;
}